requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai==2.0.10
google-generativeai==0.8.0
langchain-core==0.3.63
langchain-community
langchain-text-splitters==0.3.8
faiss-cpu
python-dotenv==1.0.1